fix: moved zoom to layer button position
