+++ sites/geohub/src/styles/button-icons-selected.scss
@@ -2,7 +2,8 @@
   opacity: 0.5;
   display: inline;
   cursor: pointer;
-  margin-right: 15px;
+  margin-left: 0.3rem;
+  margin-right: 0.3rem;
 
   &:hover {
     opacity: 1;
